Coded interface JniMatHolder in c++ in order to be able to store Mats in native memory instead of application heap, which is much smaller. This has reduced memory usage by a ton and allowed the stitching algorithm to work with the full amount of required images (Although at half dimensions. At full size, it's only possible to do around 8-9 images). Another method is now used from NativePanorama for the stitching, which is adapted to the new way of storing mats and called processFromMatHandles. Changed code in CameraFragment to reflect this. Also made it so that autoFocus is enforced whenever takePicture() is called which makes for nicer photos.
Uncommented cropping code. All in all, stitching now works as can be expected provided that there is no camera translation when taking the photos. That is, you have to rotate the phone pretty much in place. It's unfortunate, but it appears that stitching in openCV is designed for this principle of taking photos.